internal/kgateway: Move the inferencepool controller tests to it's own file
- This is consistent with the approach the base GW suite took
